requirements.txt
streamlit
scikit-learn
pandas
numpy
matplotlib
seaborn
nltk
gnews
python-dotenv
joblib
flask
flask-cors